+++ system/models.py
@@ -1,6 +1,6 @@
+from django.contrib.auth.models import User
 from django.db import models
 from django.utils import timezone
-from django.contrib.auth.models import User
 
 
 class Student(models.Model):
@@ -12,8 +12,8 @@ class Student(models.Model):
 
 class Teacher(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, default=None)
-    def __str__(self):
 
+    def __str__(self):
         return self.user.username
