+++ system/templates/system/base.html
@@ -9,7 +9,7 @@
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
 
-    <link rel="stylesheet" type="text/css" href="{% static 'main.css' %}">
+    <link rel="stylesheet" type="text/css" href="{% static 'system/main.css' %}">
 
     {% if title %}
         <title>Grade Control - {{ title }}</title>
@@ -47,9 +47,18 @@
         </div>
       </nav>
     </header>
+
     <main role="main" class="container">
       <div class="row">
-        <div class="col-md-8">
+      {% if sidebar %}
+        {% with class_name="col-md-8" %}
+        <div class="{{ class_name }}">
+        {% endwith %}
+      {% else %}
+        {% with class_name="col-md-12" %}
+        <div class="{{ class_name }}">
+        {% endwith %}
+      {% endif %}
             {% if messages %}
                 {% for message in messages %}
                     <div class="alert alert-{{ message.tags }}">
@@ -59,19 +68,21 @@
             {% endif %}
             {% block content %}{% endblock %}
         </div>
-        <!--div class="col-md-4">
-          <div class="content-section">
-            <h3>Our Sidebar</h3>
-            <p class='text-muted'>You can put any information here you'd like.
-              <ul class="list-group">
-                <li class="list-group-item list-group-item-light">Latest Posts</li>
-                <li class="list-group-item list-group-item-light">Announcements</li>
-                <li class="list-group-item list-group-item-light">Calendars</li>
-                <li class="list-group-item list-group-item-light">etc</li>
-              </ul>
-            </p>
-          </div>
-        </div-->
+        {% if sidebar %}
+            <div class="col-md-4">
+              <div class="content-section">
+                <h3>Our Sidebar</h3>
+                <p class='text-muted'>You can put any information here you'd like.
+                  <ul class="list-group">
+                    <li class="list-group-item list-group-item-light">Latest Posts</li>
+                    <li class="list-group-item list-group-item-light">Announcements</li>
+                    <li class="list-group-item list-group-item-light">Calendars</li>
+                    <li class="list-group-item list-group-item-light">etc</li>
+                  </ul>
+                </p>
+              </div>
+            </div>
+        {% endif %}
       </div>
     </main>
