+++ system/templates/system/profile.html
@@ -1,5 +1,12 @@
 {% extends "system/base.html" %}
 <!--{% load crispy_forms_tags %}-->
 {% block content %}
-   <h1>{{ user.username }}</h1>
+
+   <div class="article-metadata">
+        <h1>{{ user.username }}</h1>
+    </div>
+    <div class="media-body">
+        <p class="article-content">Welcome to your grade control panel. You are logged in as <b>{{ user.username }}</b>.</p>
+    </div>
+
 {% endblock content %}
